+++ .eslintrc.yml
@@ -34,6 +34,8 @@ rules:
   prettier/prettier: 'error'
   react/prop-types: 0
   react/react-in-jsx-scope: 'off'
+  react-hooks/rules-of-hooks: 'error' #Checks rules of Hooks
+  react-hooks/exhaustive-deps: 'warn' #Checks effect dependencies
 parser: '@typescript-eslint/parser'
 parserOptions:
   ecmaVersion: 11
