+++ package.json
@@ -109,6 +109,7 @@
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "gatsby-cli": "^2.13.0",
     "gh-pages": "^3.1.0",
     "gray-matter": "4.0.2",
