+++ src/components/AnchorTag.tsx
@@ -1,14 +1,11 @@
 import * as React from 'react';
 
-export const AnchorTag = ({ children: link, ...props }) => {
-  if (link) {
-    return (
-      <a href={props.href} target="_blank" rel="noopener noreferrer">
-        {link}
-      </a>
-    );
-  }
-  return null;
+export const AnchorTag = ({ children: link, ...props }): JSX.Element => {
+  return (
+    <a href={props.href} target="_blank" rel="noopener noreferrer">
+      {link}
+    </a>
+  );
 };
 
 export default AnchorTag;
