+++ src/components/Layout.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { MDXProvider } from '@mdx-js/react';
 import { graphql, useStaticQuery } from 'gatsby';
-import { once } from 'lodash';
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { DarkTheme } from './styles/DarkTheme';
 import mdxComponents from './styles/StyledProp';
@@ -10,92 +9,90 @@ import { ILayoutProps } from '../types/interfaces';
 import { LeftSidebar } from './LeftSidebar';
 import { Wrapper, LeftSideBarWidth, Content, MaxWidth, RightSideBarWidth } from './styles';
 
-const getSideBarData = once(
-  (): ISidebar => {
-    const { allMdx } = useStaticQuery(
-      graphql`
-        query GetNewSidebarLayoutQuery {
-          allMdx(
-            filter: { fields: { slug: { ne: "/" } } }
-            sort: { fields: fields___date, order: DESC }
-          ) {
-            group(field: fields___year) {
-              edges {
-                node {
-                  fields {
-                    date
-                    description
-                    id
-                    img
-                    slug
-                    subtitle
-                    tags
-                    title
-                    year
-                  }
+const getSideBarData = (): ISidebar => {
+  const { allMdx } = useStaticQuery(
+    graphql`
+      query GetNewSidebarLayoutQuery {
+        allMdx(
+          filter: { fields: { slug: { ne: "/" } } }
+          sort: { fields: fields___date, order: DESC }
+        ) {
+          group(field: fields___year) {
+            edges {
+              node {
+                fields {
+                  date
+                  description
+                  id
+                  img
+                  slug
+                  subtitle
+                  tags
+                  title
+                  year
                 }
               }
-              fieldValue
             }
+            fieldValue
           }
         }
-      `,
-    );
-    const allArticles = allMdx.group.sort((a, b) => {
-      switch (a.fieldValue.localeCompare(b.fieldValue, 'en', { numeric: true })) {
-        case -1:
-          return 1;
-        case 1:
-          return -1;
-        default:
-          return 0;
       }
-    });
-    const flatten = (g) => {
-      if (!g || !g.edges) return;
-      return {
-        year: g.fieldValue,
-        open: false,
-        active: false,
-        posts: g.edges.map((e) => {
-          return {
-            date: new Date(e.node.fields.date),
-            description: e.node.fields.description,
-            id: e.node.fields.id,
-            img: e.node.fields.img,
-            slug: e.node.fields.slug,
-            subtitle: e.node.fields.subtitle,
-            tags: e.node.fields.tags,
-            title: e.node.fields.title,
-            year: e.node.fields.year,
-            active: false,
-            open: false,
-          };
-        }),
-      };
-    };
-    const recent: ISection = {
-      active: true,
-      articles: allArticles.slice(0, 2).map(flatten),
-      icon: ['fas', 'mug-hot'],
-      name: 'Recent',
-      open: true,
-    };
-
-    const archive: ISection = {
-      active: false,
-      articles: allArticles.slice(3, allArticles.length - 1).map(flatten),
-      icon: ['fas', 'treasure-chest'],
-      name: 'Alcoves',
-      open: false,
-    };
-
+    `,
+  );
+  const allArticles = allMdx.group.sort((a, b) => {
+    switch (a.fieldValue.localeCompare(b.fieldValue, 'en', { numeric: true })) {
+      case -1:
+        return 1;
+      case 1:
+        return -1;
+      default:
+        return 0;
+    }
+  });
+  const flatten = (g) => {
+    if (!g || !g.edges) return;
     return {
-      sections: [recent, archive],
-      open: true,
+      year: g.fieldValue,
+      open: false,
+      active: false,
+      posts: g.edges.map((e) => {
+        return {
+          date: new Date(e.node.fields.date),
+          description: e.node.fields.description,
+          id: e.node.fields.id,
+          img: e.node.fields.img,
+          slug: e.node.fields.slug,
+          subtitle: e.node.fields.subtitle,
+          tags: e.node.fields.tags,
+          title: e.node.fields.title,
+          year: e.node.fields.year,
+          active: false,
+          open: false,
+        };
+      }),
     };
-  },
-);
+  };
+  const recent: ISection = {
+    active: true,
+    articles: allArticles.slice(0, 2).map(flatten),
+    icon: ['fas', 'mug-hot'],
+    name: 'Recent',
+    open: true,
+  };
+
+  const archive: ISection = {
+    active: false,
+    articles: allArticles.slice(3, allArticles.length - 1).map(flatten),
+    icon: ['fas', 'treasure-chest'],
+    name: 'Alcoves',
+    open: false,
+  };
+
+  return {
+    sections: [recent, archive],
+    open: true,
+  };
+};
 
 interface ISection {
   active: boolean;
@@ -111,8 +108,6 @@ interface ISidebar {
 }
 
 export const Layout: React.FC<ILayoutProps> = ({ children, location }): JSX.Element => {
-  const [collapsed] = useState(false);
-  const [toggled, setToggled] = useState(false);
   const [sidebar, setSidebar] = useState(getSideBarData());
 
   return (
@@ -120,13 +115,7 @@ export const Layout: React.FC<ILayoutProps> = ({ children, location }): JSX.Elem
       <MDXProvider components={mdxComponents}>
         <Wrapper>
           <LeftSideBarWidth className={'hiddenMobile'}>
-            <LeftSidebar
-              sidebar={sidebar}
-              collapsed={collapsed}
-              toggled={toggled}
-              handleToggleSidebar={setToggled}
-              setSidebar={setSidebar}
-            />
+            <LeftSidebar sidebar={sidebar} setSidebar={setSidebar} />
           </LeftSideBarWidth>
           <Content>
             <MaxWidth>{children}</MaxWidth>
